+++ packages/adapter-vercel/index.js
@@ -93,7 +93,6 @@ export default function ({ external = [], edge, split } = {}) {
 
 			const tmp = builder.getBuildDirectory('vercel-tmp');
 
-			builder.rimraf(dir);
 			builder.rimraf(tmp);
 
 			const files = fileURLToPath(new URL('./files', import.meta.url).href);
@@ -237,7 +236,6 @@ export default function ({ external = [], edge, split } = {}) {
 
 			builder.log.minor('Copying assets...');
 
-			builder.writeStatic(dirs.static);
 			builder.writeClient(dirs.static);
 			builder.writePrerendered(dirs.static);
 
@@ -295,6 +293,8 @@ function get_node_version() {
  * @param {string} runtime
  */
 async function create_function_bundle(builder, entry, dir, runtime) {
+	fs.rmSync(dir, { force: true, recursive: true });
+
 	let base = entry;
 	while (base !== (base = path.dirname(base)));