+++ packages/create-svelte/templates/default/src/routes/about.svelte
@@ -27,7 +27,7 @@
 		following into your command line and following the prompts:
 	</p>
 
-	<pre>npm init svelte</pre>
+	<pre>npm create svelte@latest</pre>
 
 	<p>
 		The page you're looking at is purely static HTML, with no client-side interactivity needed.