+++ packages/kit/src/node/polyfills.js
@@ -1,5 +1,7 @@
 import { fetch, Response, Request, Headers } from 'undici';
 import { ReadableStream, TransformStream, WritableStream } from 'stream/web';
+import { Readable } from 'stream';
+import { Request as NodeFetchRequest } from 'node-fetch';
 import { webcrypto as crypto } from 'crypto';
 
 /** @type {Record<string, any>} */
@@ -7,7 +9,18 @@ const globals = {
 	crypto,
 	fetch,
 	Response,
-	Request,
+	// TODO remove the superclass as soon as Undici supports formData
+	// https://github.com/nodejs/undici/issues/974
+	Request: class extends Request {
+		// @ts-expect-error
+		formData() {
+			return new NodeFetchRequest(this.url, {
+				method: this.method,
+				headers: this.headers,
+				body: this.body && Readable.from(this.body)
+			}).formData();
+		}
+	},
 	Headers,
 	ReadableStream,
 	TransformStream,