+++ packages/kit/test/apps/basics/src/routes/cyclical-dynamic-import/index.svelte
@@ -1,0 +1,8 @@
+<script>
+	import { onMount } from 'svelte';
+
+	onMount(async () => {
+		const { is_even } = await import('./_is_even.js');
+		console.log(is_even(5));
+	});
+</script>