+++ packages/kit/test/apps/basics/vite.config.js
@@ -16,8 +16,7 @@ const config = {
 	server: {
 		fs: {
 			allow: [path.resolve('../../../src')]
-		},
-		hmr: false
+		}
 	}
 };