+++ packages/kit/test/apps/options/vite.config.js
@@ -11,8 +11,7 @@ const config = {
 	server: {
 		fs: {
 			allow: [path.resolve('../../../src')]
-		},
-		hmr: false
+		}
 	}
 };