+++ packages/kit/test/apps/writes/src/routes/new-route/index.svelte
@@ -1,0 +1,1 @@
+<h1>Great success!</h1>