+++ packages/kit/test/prerendering/basics/src/routes/fetch-image.svelte
@@ -1,0 +1,2 @@
+<img src="/fetch-image/image.jpg" alt="" />
+<img src="/fetch-image/image.png" alt="" />