+++ packages/adapter-auto/CHANGELOG.md
@@ -1,5 +1,22 @@
 # @sveltejs/adapter-auto
 
+## 1.0.0-next.63
+
+### Patch Changes
+
+- Updated dependencies [[`81917206`](https://github.com/sveltejs/kit/commit/819172065cdda6c695f3fb48e60fa451c1caaae6)]:
+  - @sveltejs/adapter-cloudflare@1.0.0-next.30
+  - @sveltejs/adapter-netlify@1.0.0-next.70
+  - @sveltejs/adapter-vercel@1.0.0-next.65
+
+## 1.0.0-next.62
+
+### Patch Changes
+
+- Updated dependencies [[`2a9c1925`](https://github.com/sveltejs/kit/commit/2a9c19252912d61bd840441fe80dc13366af41f7)]:
+  - @sveltejs/adapter-vercel@1.0.0-next.64
+  - @sveltejs/adapter-netlify@1.0.0-next.69
+
 ## 1.0.0-next.61
 
 ### Patch Changes